// src/open-api/model/associatedGroup.ts
/**
 * Receipt Wrangler API.
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type AssociatedGroup = 'MINE' | 'ALL';

export const AssociatedGroup = {
    Mine: 'MINE' as AssociatedGroup,
    All: 'ALL' as AssociatedGroup
};