// src/open-api/model/sortDirection.ts
/**
 * Receipt Wrangler API.
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 5.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type SortDirection = 'asc' | 'desc' | '';

export const SortDirection = {
    Asc: 'asc' as SortDirection,
    Desc: 'desc' as SortDirection,
    Empty: '' as SortDirection
};